+++ packages/terra-folder-tree/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Changed
+  * Updated the radio button hover style to be distinct from row hover.
+
 ## 1.3.0 - (April 23, 2024)
 
 * Changed